fusion_platform/requirements.txt
scikit-learn==0.19.2
filterpy==1.4.5
numba==0.43.1
matplotlib==2.2.3
pillow==6.2.2
opencv-python==3.4.3.18
glob2==0.6
llvmlite==0.31.0 (for python 2.7)
numba==0.43.1